Fix access and clonability of the IR.
The IR has to be entirely public and since the code generator
ONLY gets a reference to it, we have to be able to copy all its
data. I'm doing that field by field at the moment because the design
includes both (useless) accessors as well as constuctors we don't
need.

The crate containing the type should have NEITHER.
An IR is a data type. There are no method of any kind
associated with it, UNLESS they're required to protect
invariants during access, for example, array bounds.
But Rust debug mode already does that.

The constraint kind (Boundary or periodic) and the segment
it applies too are candidates for protection (so you cannot
access the wrong part of a trace with a constraint) but I dount
that's worth the effort.