Fix airscript to use exemptions = 1.
Original version fixed it at 2 which is incorrect.
It meant the boundary row inspected in our 8 row test case
was row 6, instead of row 7, and the transition constraints
were only applied to rows 0->1 to 5->6 instead of all the
way to 6->7.

Unfortunately now the Cairo boundary polynomial combination
no longer agrees with Winterfell.

However the divisors do agree, so the error is in the degree
calculation but i have no idea at all what it maans so
cannot fix it.